Updated code related to calls to removed code related to redundant tables "workbench_project_user_role" and "workbench_role"
BMS-187 / BMS-4631

+++ src/main/java/org/generationcp/ibpworkbench/ui/dashboard/listener/LaunchProgramAction.java
@@ -132,7 +132,7 @@ public class LaunchProgramAction implements ItemClickListener, ClickListener {
 		try {
 
 			final ProjectUserInfoDAO projectUserInfoDao = this.workbenchDataManager.getProjectUserInfoDao();
-			final ProjectUserInfo projectUserInfo = projectUserInfoDao.getByProjectIdAndUserId(project.getProjectId().intValue(),
+			final ProjectUserInfo projectUserInfo = projectUserInfoDao.getByProjectIdAndUserId(project.getProjectId(),
 					this.sessionData.getUserData().getUserid());
 			if (projectUserInfo != null) {
 				projectUserInfo.setLastOpenDate(new Date());
